fixed some mistakes
