+++ Tools/Cmp/Regex_Engine.py
@@ -1,30 +1,30 @@
 from typing import Callable, Dict
-from Own_token import Character, Op, Token, Token_Type
+from Own_token import Token, Token_Type
 from automaton import *
 from nonr_parser import non_recursive_parse
 from regex_grammar import *
 
 regex_token_builder: Dict[str, Callable] = {
-    '*': Op('*', Token_Type.closure, 3),
-    '(': Op('(', Token_Type.open_parenthesis, 3),
-    ')': Op(')', Token_Type.closed_parenthesis, 3),
-    '|': Op('|', Token_Type.union, 1),
-    'ε': Op('ε', Token_Type.repsilon, 2),
-    '$': Op('$' ,Token_Type.eof, 1)
+    '*': Token('*', Token_Type.closure),
+    '(': Token('(', Token_Type.open_parenthesis),
+    ')': Token(')', Token_Type.closed_parenthesis),
+    '|': Token('|', Token_Type.union),
+    'ε': Token('ε', Token_Type.repsilon),
+    '$': Token('$' ,Token_Type.eof)
 }
 
 class Regex_Engine:
 
     def __init__(self, regex: str):
         self.regex = regex
-        self.automaton: 'DFA' = self.build_automaton(regex)
+        self.automaton: 'DFA' = self.build_automaton()
 
     def __call__(self, text):
         return self.automaton.match(text)
 
     
-    def build_automaton(self, regex) -> DFA:
-        tokens = self.regexTokenizer(regex)
+    def build_automaton(self) -> DFA:
+        tokens = self.regexTokenizer()
         ast, _ = non_recursive_parse(regex_grammar, tokens)
         nfa = ast.eval()
         
@@ -53,7 +53,7 @@ class Regex_Engine:
                     tokens.append(regex_token_builder[symbol])
 
                 else:
-                    tokens.append(Character(symbol, tkn_type=Token_Type.character))
+                    tokens.append(Token(symbol, token_type=Token_Type.character))
 
         return tokens
