+++ Tools/Cmp/tools.py
@@ -1,0 +1,1 @@
+EPSILON = ''
